Day 1, part A solution verified.
Prompted Copilot with the comments at the top and used
Ctrl-Enter to generate ~10 solutions.  The selected one I used
was augmented to support specifying the input filename on the
command line.  Also refactored the "get the last digit" section to
reverse search the string, rather than starting at the beginning
and traversing all the way to the end.

Go syntax errors were found related to the byte variable `c`,
because `unicode.IsDigit` and assignment to `firstDigit` and
`lastDigit` need it to be a rune.