Day 1, part B
Refactored to include looking for English words representing digits.
Had to get pretty specific and iterative on prompting to get Copilot
to do the right thing.  It tended to look for digit characters or words,
but not both at the same time.  It also was assuming the line string
was a sentence, with whitespace, and not checking for words at
every character position.